+++ runtime/src/pallets/mod.rs
@@ -12,3 +12,6 @@ pub use balances::*;
 
 pub mod transaction_payment;
 pub use transaction_payment::*;
+
+pub mod parachain_system;
+pub use parachain_system::*;
